pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120

[project]
name = "codecritic-ai"
version = "0.1.0"
description = "CodeCritic AI — an AI-powered platform for large-scale code analysis using RAG, vector search, and distributed pipelines."
readme = "README.md"
requires-python = ">=3.9,<=3.11"
authors = [
    { name = "Pranav Nagothu", email = "pnagothu@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "GitPython==3.1.43",
    "Pygments==2.18.0",
    "anytree==2.12.1",
    "cohere==5.9.2",
    "configargparse",
    "faiss-cpu==1.9.0",
    "fastapi==0.112.2",
    "google-ai-generativelanguage==0.6.6",
    "gradio>=4.26.0",
    "langchain==0.2.16",
    "langchain-anthropic==0.1.23",
    "langchain-cohere==0.2.4",
    "langchain-community==0.2.17",
    "langchain-core==0.2.41",
    "langchain-experimental==0.0.65",
    "langchain-google-genai",
    "langchain-nvidia-ai-endpoints==0.2.2",
    "langchain-ollama==0.1.3",
    "langchain-openai==0.1.25",
    "langchain-text-splitters==0.2.4",
    "langchain-voyageai==0.1.1",
    "langchain-milvus==0.1.6",
    "langchain-chroma==0.1.4",
    "langchain-qdrant==0.1.4",
    "marqo==3.7.0",
    "nbformat==5.10.4",
    "openai==1.42.0",
    "pinecone==5.0.1",
    "pinecone-text==0.9.0",
    "python-dotenv==1.0.1",
    "python-Levenshtein==0.26.0",
    "requests==2.32.3",
    "semchunk==2.2.0",
    "sentence-transformers==3.1.0",
    "tenacity==8.5.0",
    "tiktoken==0.7.0",
    "tokenizers==0.19.1",
    "transformers==4.44.2",
    "tree-sitter==0.22.3",
    "tree-sitter-typescript==0.21.2",
    "tree-sitter-language-pack==0.2.0",
    "voyageai==0.2.3",
    "setuptools"  # Added from the setup.py install_requires
]

[project.optional-dependencies]
dev = ["black"]

[project.scripts]
codecritic-index = "code_chatbot.ingestion.indexer:main"
codecritic-chat = "code_chatbot.retrieval.rag:main"

[tool.setuptools.package-data]
"code_chatbot" = ["sample-exclude.txt"]

[tool.setuptools.packages]
find = {}
